refactor(bit-remove, bit-merge-lane): extract bit-remove and bit-merge-lane to new aspects
