+++ scopes/component/remove/index.ts
@@ -1,0 +1,5 @@
+import { RemoveAspect } from './remove.aspect';
+
+export type { RemoveMain } from './remove.main.runtime';
+export default RemoveAspect;
+export { RemoveAspect };
