+++ scopes/component/remove/remove.aspect.ts
@@ -1,0 +1,5 @@
+import { Aspect } from '@teambit/harmony';
+
+export const RemoveAspect = Aspect.create({
+  id: 'teambit.component/remove',
+});
