+++ scopes/lanes/lanes/lanes.main.runtime.ts
@@ -44,7 +44,6 @@ import {
 } from './lane.cmd';
 import { lanesSchema } from './lanes.graphql';
 import { SwitchCmd } from './switch.cmd';
-import { mergeLanes } from './merge-lanes';
 import { LaneSwitcher } from './switch-lanes';
 import { createLane, throwForInvalidLaneName } from './create-lane';
 
@@ -53,20 +52,6 @@ export { Lane };
 export type LaneResults = {
   lanes: LaneData[];
   currentLane?: string | null;
-};
-
-export type MergeLaneOptions = {
-  mergeStrategy: MergeStrategy;
-  noSnap: boolean;
-  snapMessage: string;
-  existingOnWorkspaceOnly: boolean;
-  build: boolean;
-  keepReadme: boolean;
-  squash: boolean;
-  pattern?: string;
-  includeDeps?: boolean;
-  skipDependencyInstallation?: boolean;
-  remote?: boolean;
 };
 
 export type CreateLaneOptions = {
@@ -374,25 +359,6 @@ export class LanesMain {
     if (this.workspace) await this.workspace.consumer.onDestroy();
 
     return results.laneResults;
-  }
-
-  async mergeLane(
-    laneName: string,
-    options: MergeLaneOptions
-  ): Promise<{ mergeResults: ApplyVersionResults; deleteResults: any }> {
-    if (!this.workspace) {
-      throw new BitError(`unable to merge a lane outside of Bit workspace`);
-    }
-    const results = await mergeLanes({
-      merging: this.merging,
-      workspace: this.workspace,
-      lanesMain: this,
-      laneName,
-      ...options,
-    });
-
-    await this.workspace.consumer.onDestroy();
-    return results;
   }
 
   /**
