+++ scopes/lanes/merge-lanes/index.ts
@@ -1,0 +1,5 @@
+import { MergeLanesAspect } from './merge-lanes.aspect';
+
+export type { MergeLanesMain } from './merge-lanes.main.runtime';
+export default MergeLanesAspect;
+export { MergeLanesAspect };
