+++ scopes/lanes/merge-lanes/merge-lanes.aspect.ts
@@ -1,0 +1,5 @@
+import { Aspect } from '@teambit/harmony';
+
+export const MergeLanesAspect = Aspect.create({
+  id: 'teambit.lanes/merge-lanes',
+});
